# openapi.yaml
openapi: 3.1.0
info:
  title: SYNTARO API — SYNTARO
  description: |
    SYNTARO is an open-source GitHub bot that turns labeled issues into pull requests.

    ## RapidAPI Marketplace
    
    SYNTARO is also available as a payable API on the [RapidAPI Marketplace](https://rapidapi.com/).
    Subscribe to a plan to get an API key and start submitting fix jobs programmatically.
    
    Endpoints marked with the `RapidAPI` tag require a valid RapidAPI subscription.
    Use `X-RapidAPI-Key` header for authentication and `X-RapidAPI-Proxy-Secret` for proxy validation.
    
    ### Rate Limiting (RapidAPI)
    
    RapidAPI subscribers are rate-limited by plan tier:
    - **Free**: 10 requests/day
    - **Pro**: 100 requests/day
    - **Enterprise**: 1000 requests/day
    
    Rate limit headers are returned with every response:
    - `X-RateLimit-Limit` — Maximum requests per window
    - `X-RateLimit-Remaining` — Requests remaining in the current window
    - `X-RateLimit-Plan` — Your current plan tier
    - `Retry-After` — Seconds to wait before retrying (only on 429 responses)
    
    ### Example Curl
    
    ```bash
    # Submit a fix job
    curl -X POST https://syntaro-rapidapi.p.rapidapi.com/api/fix \
      -H "Content-Type: application/json" \
      -H "X-RapidAPI-Key: your-rapidapi-key" \
      -H "X-RapidAPI-Proxy-Secret: your-proxy-secret" \
...
    
    # Poll job status
    curl https://syntaro-rapidapi.p.rapidapi.com/api/fix/<jobId> \
...
    ```
...

    ## Authentication
...
servers:
...
paths:
...
  /api/v1/credits/usage:
    get:
      tags:
        - Credits
      summary: Credit usage statistics
      description: |
        Returns credit consumption statistics for the authenticated account,
        aggregated by day, week, or month. Only deduction transactions
        (negative amounts) are included.
      operationId: getCreditUsage
      security:
        - BearerToken: []
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [daily, weekly, monthly]
            default: monthly
          description: Aggregation period
      responses:
        "200":
          description: Usage statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditUsageResponse"
              example:
                accountId: 42
                period: "monthly"
                usage:
                  - periodStart: "2026-01-01T00:00:00.000Z"
                    totalCredits: 150
                    totalTransactions: 3
                  - periodStart: "2026-02-01T00:00:00.000Z"
                    totalCredits: 200
                    totalTransactions: 4
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ──────────────────────────────────────────────────────────────────────────
  # Admin — Credit Adjust
  # ──────────────────────────────────────────────────────────────────────────
...
  /api/v1/me/usage:
    get:
      tags:
        - User
      summary: Current user usage
      description: Returns usage statistics for the authenticated user.
      operationId: getMeUsage
      security:
        - BearerToken: []
      responses:
        "200":
          description: User usage data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserUsage"
              example:
                totalRuns: 42
                totalCreditsUsed: 2100
                monthlyCreditsUsed: 450
                monthlyRuns: 9
                freeTierRemaining: 55
                tier: "pro"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    RapidAPIKey:
      type: apiKey
      in: header
      name: X-RapidAPI-Key
      description: |
        API key for the RapidAPI Marketplace. Subscribe to a plan at
        https://rapidapi.com/ to get your API key.
        
        Pass the key in the `X-RapidAPI-Key` header with every request to
        RapidAPI-protected endpoints.
        
        **Rate limit headers** are returned with every response:
        - `X-RateLimit-Limit` — Maximum requests per window (based on plan)
        - `X-RateLimit-Remaining` — Requests remaining in current window
        - `X-RateLimit-Plan` — Your current plan tier
    
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        API key or user token provided in the Authorization header.
    AdminKey:
      type: apiKey
      in: header
      name: x-admin-key
      description: |
        Admin API key provided in the x-admin-key header.
        Must match the configured ADMIN_API_KEY environment variable.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: string
          description: Detailed error information (optional)
      required:
        - error

    # ── Webhook ────────────────────────────────────────────────────────────
...
    CreditUsageResponse:
      type: object
      properties:
        accountId:
          type: integer
          description: The authenticated account ID
        period:
          type: string
          enum: [daily, weekly, monthly]
          description: Aggregation period
        usage:
          type: array
          items:
            $ref: "#/components/schemas/UsagePeriod"
      required:
        - accountId
        - period
        - usage
    UsagePeriod:
      type: object
      properties:
        periodStart:
          type: string
          format: date-time
          description: Start of the aggregation period
        totalCredits:
          type: integer
          description: Total credits consumed in this period
        totalTransactions:
          type: integer
          description: Number of deduction transactions in this period
      required:
        - periodStart
        - totalCredits
        - totalTransactions
...
    UserUsage:
      type: object
      properties:
        totalRuns:
          type: integer
        totalCreditsUsed:
          type: integer
        monthlyCreditsUsed:
          type: integer
        monthlyRuns:
          type: integer
        freeTierRemaining:
          type: integer
          nullable: true
        tier:
          type: string
      required:
        - totalRuns
        - totalCreditsUsed
        - tier

    # ── Admin ───────────────────────────────────────────────────────────────
...
  responses:
...
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...